% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{BindSARS}
\alias{BindSARS}
\title{Bind a List of SARS Objects}
\usage{
BindSARS(sars_list, ids = NULL)
}
\arguments{
\item{sars_list}{a list of SARS object to be combined}

\item{ids}{a numeric vector (not a list!) for id of each SARS object}
}
\value{
a SARS object with an extra id vector as attributes
}
\description{
Bind a List of SARS Objects
}
\examples{
states <- matrix(c(1, 2, 3, 4), 2, 2)
actions <- matrix(c(1, 0), 2, 1)
rewards <- matrix(c(1, 2), 2, 1)
states_next <- matrix(c(2, 3, 4, 5), 2, 2)
ss <- SARS(states, actions, rewards, states_next)
BindSARS(list(ss, ss, ss))
}
