% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cartpole.R
\name{CartPole}
\alias{CartPole}
\title{CartPole Object}
\usage{
CartPole(
  internal_state = vector("numeric", 4),
  rng_state = NULL,
  setting = list()
)
}
\arguments{
\item{internal_state}{vector of size 4, representing cart position, cart velocity,
pole angle, and pole angular velocity}

\item{rng_state}{the RNG state for non-deterministic evolution}

\item{setting}{a list of constants that controls the setting of cart-pole environment,
see details.}
}
\value{
a CartPole object (\code{class = c("CartPole", "Env")})
}
\description{
The function \code{CartPole()} creates an CartPole object. The \code{CartPole} class is
inherited from the \code{Env} class.
}
\details{
blah blah
}
\examples{
cp <- CartPole()
}
