% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{CumReward}
\alias{CumReward}
\title{Cumulative Reward}
\usage{
CumReward(rewards, discount)
}
\arguments{
\item{rewards}{a numeric vector of reward sequence}

\item{discount}{a numeric number between 0 and 1}
}
\value{
a numeric number for cumulative reward
}
\description{
\code{CumReward()} is a utility function that computes cumulative reward for discounted
MDP.
}
\examples{
CumReward(c(1, 1, 0, 1), 0.9)
}
