% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.logsumexp}
\alias{.logsumexp}
\title{LogSumExp}
\usage{
.logsumexp(x)
}
\arguments{
\item{x}{a numeric object}
}
\value{
LogSumExp of x
}
\description{
\code{.logsumexp()} is an internal function to compute LogSumExp (aka SoftMax), using
the shifting trick to increase numerical stability.
}
\examples{
.logsumexp(1:3)
}
