% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{Env}
\alias{Env}
\title{Environment Object}
\usage{
Env(internal_state = NULL, rng_state = .Random.seed)
}
\arguments{
\item{internal_state}{the internal state ("everything is Markov")}

\item{rng_state}{the RNG state for non-deterministic evolution}
}
\value{
an environment object (\code{class = "env"})
}
\description{
Environment Object
}
\note{
This is an abstract class, and should not be used directly.
}
