% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{EpsilonGreedy}
\alias{EpsilonGreedy}
\title{Epsilon-Greedy Action}
\usage{
EpsilonGreedy(values, epsilon = 0.1)
}
\arguments{
\item{values}{a numeric vector or matrix for values of actions. For numeric vector,
it is values of all actions; for numeric matrix, each row is a set of values.}

\item{epsilon}{a numeric value between 0 and 1. It is the probability of choosing
uniformly random actions instead of greedy actions.}
}
\value{
an integer vector for chosen action indexes
}
\description{
\code{EpsilonGreedy()} is a utility function that performs "epsilon-greedy" actions.
}
\examples{
EpsilonGreedy(c(2, 1, 3, 7, 5))
EpsilonGreedy(matrix(c(1, 2, 2, 1), 2, 2, byrow = TRUE), epsilon = 0.2)
}
