% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Gibbs}
\alias{Gibbs}
\title{Gibbs Action}
\usage{
Gibbs(values, temperature = 1)
}
\arguments{
\item{values}{a numeric vector or matrix for values of actions. For numeric vector,
it is values of all actions; for numeric matrix, each row is a set of values.}

\item{temperature}{a numeric value as temperature.}
}
\value{
an integer vector for chosen action indexes
}
\description{
\code{Gibbs()} is a utility function that samples actions from Gibbs distribution.
}
\details{
The Gibbs (or Boltzmann) distribution has p.d.f. of the following form
\deqn{p(x) \propto \exp(x / T),}{p(x) = C exp(x / T),}
where temperature \eqn{T} controls the trade-off between greedy (exploitation)
and uniformly random (exploration): a high temperature makes the distribution
more even, while a lower temperature makes the distribution more concentrated
(to favor higher values).
}
\note{
A negative \eqn{T} can be used to favor actions with lower values. In fact, the
standard formulation of Gibbs distribution has a negative sign before \eqn{x / T},
thus lower values have higher chances of being sampled.
}
\examples{
Gibbs(c(2, 1, 3, 7, 5))
Gibbs(matrix(c(1, 2, 2, 1), 2, 2, byrow = TRUE), temperature = 0.01)
}
