% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn.R
\name{GradientFQI}
\alias{GradientFQI}
\alias{GradientGGQ}
\alias{GradientBEM}
\title{Gradient functions for Q-learning}
\usage{
GradientFQI(theta, phis, discount)

GradientGGQ(theta, phis, discount)

GradientBEM(theta, phis, discount)
}
\arguments{
\item{theta}{a numeric vector as model parameter.}

\item{phis}{a list of processed outcome from \code{SARS2Phis()}.}

\item{discount}{a numeric number between 0 and 1.}
}
\value{
gradient
}
\description{
Gradient functions for Q-learning
}
\section{Functions}{
\itemize{
\item \code{GradientFQI}: Fitted Q Iteration (objective function: MSPBE)

\item \code{GradientGGQ}: Greedy Gradient-Q (objective function: MSPBE)

\item \code{GradientBEM}: Bellman Error Minimization (objective function: MSBE)
}}

