% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{greedy}
\alias{greedy}
\title{Greedy Action}
\usage{
greedy(values)
}
\arguments{
\item{values}{a numeric vector or matrix for values of actions. For numeric vector,
returns the index of the maximum; for numeric matrix, returns the indexes of the
maximums for each row.}
}
\value{
an integer vector
}
\description{
The function \code{greedy()} is a utility function that returns action indexes of maximal
values.
}
\examples{
greedy(c(2, 1, 3, 7, 5))
greedy(matrix(c(1, 2, 2, 1), 2, 2, byrow = TRUE))
}
