% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{LogSumExp}
\alias{LogSumExp}
\title{LogSumExp}
\usage{
LogSumExp(x)
}
\arguments{
\item{x}{a numeric object}
}
\value{
LogSumExp of x
}
\description{
\code{LogSumExp()} is a utility function to compute LogSumExp (aka SoftMax), using
the shifting trick to increase numerical stability.
}
\examples{
LogSumExp(1:3)
}
