% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn.R
\name{MSPBE}
\alias{MSPBE}
\alias{MSBE}
\title{Objective functions for Q-learning}
\usage{
MSPBE(theta, phis, discount)

MSBE(theta, phis, discount)
}
\arguments{
\item{theta}{a numeric vector as model parameter.}

\item{phis}{a list of processed outcome from \code{SARS2Phis()}.}

\item{discount}{a numeric number between 0 and 1.}
}
\value{
a (non-negative) number
}
\description{
Objective functions for Q-learning
}
\section{Functions}{
\itemize{
\item \code{MSPBE}: Mean Squared Projected Bellman Error

\item \code{MSBE}: Mean Squared Bellman Error
}}

