% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{Observe}
\alias{Observe}
\alias{Observe.Env}
\title{Observe from Environment}
\usage{
Observe(x)

\method{Observe}{Env}(x)
}
\arguments{
\item{x}{an environment-like object}
}
\value{
an observation
}
\description{
\code{Observe()} is a generic method that aims to extract an observation from an object.
}
\section{Methods (by class)}{
\itemize{
\item \code{Env}: Internal state from an environment object
}}

\examples{
ev <- Env(1)
Observe(ev)
}
