% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Poly}
\alias{Poly}
\title{Polynomial Basis}
\usage{
Poly(x, degree = 1, interaction_only = TRUE, include_bias = TRUE)
}
\arguments{
\item{x}{a data matrix, where rows are observations and columns are dimensions.}

\item{degree}{degree of polynomial.}

\item{interaction_only}{if \code{TRUE}, only interactions (products of distinct features)
are produced.}

\item{include_bias}{if \code{TRUE}, include a column of ones}
}
\value{
transformed matrix
}
\description{
\code{Poly()} is a utility function for polynomial transform.
}
\examples{
x <- cbind(c(0, 1, 0, 1), 1 : 4)
Poly(x, degree = 2)
}
