% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{RBF}
\alias{RBF}
\title{Radial Basis Function}
\usage{
RBF(x, centers, Kernel = Gaussian, scale = 1, include_bias = TRUE)
}
\arguments{
\item{x}{a data matrix, where rows are observations and columns are dimensions.}

\item{centers}{a matrix for cluster centers}

\item{Kernel}{radial functions}

\item{scale}{scale parameter, or bandwidth}

\item{include_bias}{if \code{TRUE}, include a column of ones}
}
\value{
transformed matrix
}
\description{
\code{RBF()} is a utility function for RBF transform.
}
\examples{
RBF(matrix(1:12, 12, 1), matrix(c(3, 9), 2, 1))
}
\seealso{
\code{\link{Gaussian}}
}
