% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{RowWiseKronecker}
\alias{RowWiseKronecker}
\title{Row-Wise Kronecker}
\usage{
RowWiseKronecker(x, y)
}
\arguments{
\item{x}{the first data matrix}

\item{y}{the second data matrix, should have the same number of rows as \code{x}}
}
\value{
transformed matrix
}
\description{
\code{RowWiseKronecker()} is a row-wise Kronecker utility function for feature engineering.
}
\examples{
x <- cbind(rep(1, 3), 1 : 3)
y <- matrix(c(1, 0,
              0, 1,
              1, 1), 3, 2, byrow = TRUE)
 RowWiseKronecker(x, y)
}
