% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob.R
\name{RZeroInflatedPoisson}
\alias{RZeroInflatedPoisson}
\title{Generates a sample from zero-inflated Poisson}
\usage{
RZeroInflatedPoisson(n, p, lambda)
}
\arguments{
\item{n}{sample size}

\item{p}{probability of extra zeros}

\item{lambda}{rate of Poisson distribution}
}
\value{
a vector of length n
}
\description{
Generates a sample from zero-inflated Poisson
}
\examples{
RZeroInflatedPoisson(5, 0.4, 5)
}
