% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn.R
\name{SARS2Phis}
\alias{SARS2Phis}
\title{Convert SARS to Basis Representation}
\usage{
SARS2Phis(sars, action_space, Feature)
}
\arguments{
\item{sars}{a SARS object (\code{class = "SARS"}).}

\item{action_space}{a list of feasible actions.}

\item{Feature}{feature function, taking states and actions as input, outputs basis.}
}
\value{
a list containing the following components:
\describe{
\item{\code{phi}}{data matrix of \eqn{\phi(state, action)}}
\item{\code{phi_next_list}}{a list of data matrices, each matrix has the same size
as \code{phi} and represents the next time-step data matrix with each possible action}
\item{\code{r}}{rewards}
\item{\code{n}}{sample size}
}
}
\description{
The function \code{SARS2Phis()} converts a SARS object to basis representation. The
motivation is to preprocess data and prepare for faster training.
}
\examples{
states <- matrix(c(1, 2, 3, 4), 2, 2)
actions <- matrix(c(1, 0), 2, 1)
rewards <- matrix(c(1, 2), 2, 1)
states_next <- matrix(c(2, 3, 4, 5), 2, 2)
sars <- SARS(states, actions, rewards, states_next)
Feature <- function(states, actions) {
  phi <- RowWiseKronecker(Poly(states, 2), Poly(actions, 1))
  return(phi)
}
SARS2Phis(sars, list(0, 1), Feature)
}
