% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{Seed}
\alias{Seed}
\alias{Seed.Env}
\title{Set Seed for Environment}
\usage{
Seed(x, seed = NULL)

\method{Seed}{Env}(x, seed = NULL)
}
\arguments{
\item{x}{an environment-like object}

\item{seed}{an integer for \code{set.seed()}}
}
\value{
the environment-like object with seed set
}
\description{
\code{Seed()} is a generic method that aims to set seed for an object.
}
\section{Methods (by class)}{
\itemize{
\item \code{Env}: Set seed for an environment object, changing its \code{rng_state}
}}

\examples{
ev <- Env()
ev <- Seed(ev, 100)
}
