% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Soft}
\alias{Soft}
\title{Soft Thresholding}
\usage{
Soft(z, lambda = 0)
}
\arguments{
\item{z}{input numeric object}

\item{lambda}{threshold}
}
\value{
output numeric object
}
\description{
Soft Thresholding
}
\examples{
Soft(c(-1, 0, 2, 3), 1)
}
