% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{Step}
\alias{Step}
\alias{Step.Env}
\title{Evolve the Environment}
\usage{
Step(x, action)

\method{Step}{Env}(x, action)
}
\arguments{
\item{x}{an environment-like object}

\item{action}{an action from agent}
}
\value{
the environment-like object with seed set
}
\description{
\code{Step()} is a generic method that aims to evolve one step for an object.
}
\section{Methods (by class)}{
\itemize{
\item \code{Env}: Evolve one step for an environment object, changing its \code{internal_state}
}}

\examples{
ev <- Env()
ev <- Step(ev, 2e5)
}
