% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{trajectory}
\alias{trajectory}
\title{Trajectory Object}
\usage{
trajectory(observations, actions)
}
\arguments{
\item{observations}{a list for observations, the number of elements should be
one more than that of \code{actions}.}

\item{actions}{a list for actions, each element for each time step.}
}
\value{
a trajectory object (\code{class = "trajectory"})
}
\description{
The function \code{trajectory()} creates a trajectory object for discrete-time RL data.
}
\details{
The trajectory object is designed to represent typical RL trajectory data:
\deqn{O_1, A_1, O_2, A_2, \ldots, O_n, A_n, O_{n+1}}

This representation is compatible with many RL data generator (for example the
data structure of OpenAI Gym library). With user-defined interpreter, observations
and actions (or their history) can be encoded/converted to states, rewards and
actions under Markov Decision Process (MDP) framework.
}
\examples{
observations <- list(0, -1, -1, 0, 1, 1)
actions <- list(-1, 1, 1, 0, -1)
traj <- trajectory(observations, actions)
traj
}
