% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cartpole.R
\name{CartPole}
\alias{CartPole}
\alias{Step.CartPole}
\alias{Reset.CartPole}
\title{CartPole Object}
\usage{
CartPole(
  internal_state = vector("numeric", 4),
  rng_state = .Random.seed,
  setting = list()
)

\method{Step}{CartPole}(x, action)

\method{Reset}{CartPole}(x)
}
\arguments{
\item{internal_state}{vector of size 4, representing cart position, cart velocity,
pole angle, and pole angular velocity}

\item{rng_state}{the RNG state for non-deterministic evolution}

\item{setting}{a list of constants that controls the setting of the CartPole object,
see details.}

\item{x}{an environment-like object}

\item{action}{an action from agent}
}
\value{
a CartPole object (\code{class = c("CartPole", "Env")})
}
\description{
The function \code{CartPole()} creates a CartPole object. The \code{CartPole} class is inherited
from the \code{Env} class.
}
\details{
The internal state of a CartPole object includes four elements (following OpenAI
Gym convention):\tabular{ll}{
   Index \tab Observation \cr
   1 \tab Cart Position \cr
   2 \tab Cart Velocity \cr
   3 \tab Pole Angle \cr
   4 \tab Pole Angular Velocity \cr
}


Note that the \eqn{\theta} (index 3 and 4) has the opposite sign to OpenAI Gym.
In our definition, anti-clockwise is positive direction.
}
\section{Methods (by generic)}{
\itemize{
\item \code{Step}: evolve given left (-1) or right (+1) action

\item \code{Reset}: reset by uniformly sample from U[-0.05, +0.05]
}}

\examples{
cp <- CartPole()
}
