% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{Env}
\alias{Env}
\title{Environment Object}
\usage{
Env(internal_state = 0, rng_state = .Random.seed)
}
\arguments{
\item{internal_state}{the internal state ("everything is Markov")}

\item{rng_state}{the RNG state for non-deterministic evolution}
}
\value{
an environment object (\code{class = "Env"})
}
\description{
The function \code{Env()} creates an environment object.
}
\details{
This is an abstract class, and should not be used directly. The purpose of this
class is to provide a recommended template for developers. Standard methods include
\code{Observe}, \code{Seed}, \code{Step}, and \code{Reset}.
}
\examples{
Env(1)
}
