% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Greedy}
\alias{Greedy}
\title{Greedy Action}
\usage{
Greedy(values)
}
\arguments{
\item{values}{a numeric vector or matrix for values of actions. For numeric vector,
it is values of all actions; for numeric matrix, each row is a set of values.}
}
\value{
an integer vector for chosen action indexes
}
\description{
\code{Greedy()} is a utility function that returns action indexes of maximal values.
}
\examples{
Greedy(c(2, 1, 3, 7, 5))
Greedy(matrix(c(1, 2, 2, 1), 2, 2, byrow = TRUE))
}
\seealso{
Other value-based action functions: 
\code{\link{EpsilonGreedy}()},
\code{\link{Gibbs}()},
\code{\link{Random}()}
}
\concept{value-based action functions}
