% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{Observe}
\alias{Observe}
\alias{Observe.default}
\alias{Observe.Env}
\title{Observe}
\usage{
Observe(x)

\method{Observe}{default}(x)

\method{Observe}{Env}(x)
}
\arguments{
\item{x}{environment-like object}

\item{...}{extra arguments}
}
\value{
an observation
}
\description{
\code{Observe()} is a generic method that aims to extract an observation from an object.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Default behavior (error)

\item \code{Env}: Internal state from an environment object
}}

\note{
\code{Observe()} is not meant to be used directly on an environment object. For developers,
it is best practice to define an inheritance of environment class, then define
corresponding methods.
}
