% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{random}
\alias{random}
\title{Random Action}
\usage{
random(values)
}
\arguments{
\item{values}{a numeric vector or matrix for values of actions. For numeric vector,
it is values of all actions; for numeric matrix, each row is a set of values.}
}
\value{
an integer vector for chosen action indexes
}
\description{
\code{random()} is a utility function that returns randomly picked action indexes, with
probabilities determined by input values.
}
