% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{RBF}
\alias{RBF}
\title{Radial Basis Function}
\usage{
RBF(
  x,
  centers,
  Kernel = Gaussian,
  scale = 1,
  include_bias = TRUE,
  normalize = FALSE
)
}
\arguments{
\item{x}{a data matrix, where rows are observations and columns are dimensions.}

\item{centers}{a matrix for cluster centers}

\item{Kernel}{radial functions}

\item{scale}{scale parameter, or bandwidth}

\item{include_bias}{if \code{TRUE}, include a column of ones}

\item{normalize}{if \code{TRUE}, features (expect for bias) are normalized}
}
\value{
transformed matrix
}
\description{
\code{RBF()} is a utility function for RBF transform.
}
\examples{
x <- matrix(rnorm(20), 10, 2)
centers <- kmeans(x, 2)$centers
RBF(x, centers)
}
\seealso{
\code{\link{Gaussian}}
}
